% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_dly}
\alias{clean_dly}
\title{Clean dly data}
\usage{
clean_dly(data)
}
\arguments{
\item{data}{Data frame of dly data.}
}
\value{
Data frame.
}
\description{
Transforms dly data so that each row is one day, and each column is a
variable. Removes flag variables. Also attempts to make units of measurement
consistent across variables.
}
