% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_vals}
\alias{clean_vals}
\title{Make unit of measurement consistent across variables}
\usage{
clean_vals(x, element)
}
\arguments{
\item{x}{Vector of values of type integer or double}

\item{element}{GHCN element type of size x. For example, "PRCP" is the
element type for precipitation.}
}
\value{
Vector of type double.
}
\description{
Always try to use millimeters for variables that measure length, and degrees
Celsius for variables that measure temperature.
}
