% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database}
\alias{create_database}
\title{Create database of daily data}
\usage{
create_database(dir_files, file_db, chunk_size = 100)
}
\arguments{
\item{dir_files}{Directory containing the .dly data files.}

\item{file_db}{Path and file name of SQLite database to create.}

\item{chunk_size}{Number of files of data to append at a time to database
table.}
}
\description{
Create database of daily data
}
