% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database_dir}
\alias{create_database_dir}
\title{Create database of daily data from directory of ".dly" data files}
\usage{
create_database_dir(dir, file_db, chunk_size = 100)
}
\arguments{
\item{dir}{Path of directory containing ".dly" data files.}

\item{file_db}{File path of new database to write to.}

\item{chunk_size}{Number of files of data to append at a time to database
table.}
}
\description{
Create database of daily data from a directory of ".dly" data files. The
data will only include the five core statistics daily precipitation, snow
fall, snow depth, minimum temperature, and maximum temperature.
}
