% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_read_dly}
\alias{db_read_dly}
\title{Read database dly table.}
\usage{
db_read_dly(file_db, ids, vars)
}
\arguments{
\item{file_db}{Path of database file to read.}

\item{ids}{Station IDs in data to query (where).}

\item{vars}{Variables in data to query (select).}
}
\value{
Data frame.
}
\description{
Read database dly table.
}
