% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_stations}
\alias{find_stations}
\title{Find stations that match pattern}
\usage{
find_stations(pattern)
}
\arguments{
\item{pattern}{Character string to match station names.}
}
\value{
Data frame.
}
\description{
Find stations that match pattern
}
