% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-aws-helpers.R
\name{get_inventory_aws}
\alias{get_inventory_aws}
\title{Get GHCN-D Station Inventory Information for Multiple Stations}
\usage{
get_inventory_aws(station_ids)
}
\arguments{
\item{station_ids}{Character vector. One or more 11-character GHCN station identification codes.}
}
\value{
A data.table containing inventory information for the specified stations.
}
\description{
This function retrieves inventory information for specified GHCN-D station(s),
including available variables, their coverage periods, and other relevant details.
}
\examples{
# Get inventory for two stations
get_station_inventory(c("USW00094728", "USC00305816"))

}
