% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-package.R
\docType{package}
\name{ghcn-package}
\alias{ghcn}
\alias{ghcn-package}
\title{ghcn: Provides an interface to NCEI GHCN-Daly dataset.}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://joshbrinks.github.io/ghcn/}
}

}
\keyword{internal}
