% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghcn.stations}
\alias{ghcn.stations}
\title{GHCN-D Stations Data}
\format{
A data table with 9 variables:
\describe{
  \item{id}{character. 11-character station identification code.}
  \item{latitude}{numeric. Latitude of the station in decimal degrees.}
  \item{longitude}{numeric. Longitude of the station in decimal degrees.}
  \item{elevation}{numeric. Elevation of the station in meters.}
  \item{state}{character. U.S. postal code for the state (for U.S. and Canadian stations only).}
  \item{name}{character. Name of the station.}
  \item{gsn_flag}{character. Flag indicating if the station is part of the GCOS Surface Network (GSN).}
  \item{hcn_crn_flag}{character. Flag indicating if the station is part of the U.S. Historical Climatology Network (HCN) or U.S. Climate Reference Network (CRN).}
  \item{wmo_id}{character. World Meteorological Organization (WMO) number for the station.}
}
}
\source{
\url{http://noaa-ghcn-pds.s3.amazonaws.com/ghcnd-stations.txt}
}
\usage{
ghcn.stations
}
\description{
A dataset containing information about weather stations in the Global Historical Climatology Network - Daily (GHCN-D).
}
\keyword{datasets}
