% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{ghcn_clean}
\alias{ghcn_clean}
\title{Clean dly data}
\usage{
ghcn_clean(data)
}
\arguments{
\item{data}{Data frame of dly data read with `ghcn_read()`.}
}
\value{
Data frame of dly data.
}
\description{
Transforms dly data so that each row is one day, and each column is a
variable. Only includes the five core statistics daily precipitation, snow
fall, snow depth, minimum temperature, and maximum temperature. Removes flag
variables. Also makes units of measurement consistent across variables.
}
