% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghcn_daily_datatypes}
\alias{ghcn_daily_datatypes}
\title{GHCN Daily Datatypes}
\format{
A data frame with 56 rows and 2 variables:
\describe{
  \item{datatype}{GHCN daily datatype code}
  \item{description}{Description of what the datatype represents}
}
}
\source{
\url{https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}
}
\usage{
ghcn_daily_datatypes
}
\description{
A dataset containing the Global Historical Climatology Network (GHCN) daily
datatypes and their descriptions.
}
\keyword{datasets}
