% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ghcn_db_read}
\alias{ghcn_db_read}
\title{Read database dly table.}
\usage{
ghcn_db_read(file_db, ids, vars)
}
\arguments{
\item{file_db}{Path of database file to read.}

\item{ids}{Station IDs in data to query (where).}

\item{vars}{Variables in data to query (select).}
}
\value{
Data frame.
}
\description{
Read database dly table.
}
