% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ghcn_db_write}
\alias{ghcn_db_write}
\title{Create database of daily data from ".dly" data files}
\usage{
ghcn_db_write(files, file_db, table_name, batch_size = 100)
}
\arguments{
\item{files}{File paths of dly data files}

\item{file_db}{File path of database}

\item{table_name}{Database table name}

\item{batch_size}{Number of data files to write to database table per batch}
}
\description{
Create database of daily data from a vector of ".dly" data files. The data
will only include the five core statistics daily precipitation, snow
fall, snow depth, minimum temperature, and maximum temperature. If the
database already exists, it will be deleted first.
}
