% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{ghcn_find_stations}
\alias{ghcn_find_stations}
\title{Find stations that match pattern}
\usage{
ghcn_find_stations(pattern)
}
\arguments{
\item{pattern}{Character string of length one to match station names.}
}
\value{
Data frame of station data.
}
\description{
Find stations that match pattern
}
