% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ghcn_read}
\alias{ghcn_read}
\title{Read a ".dly" file}
\usage{
ghcn_read(id)
}
\arguments{
\item{id}{Station identification code of length one.}
}
\value{
Data frame of dly data.
}
\description{
Read a ".dly" file from the NOAA FTP site, given the station identification
code.
}
