% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ghcn_read_file}
\alias{ghcn_read_file}
\title{Read a ".dly" file}
\usage{
ghcn_read_file(file)
}
\arguments{
\item{file}{Path to ".dly" data file or connection.}
}
\value{
Data frame of dly data..
}
\description{
Read a ".dly" file, given the file path.
}
