% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ghcn_read_stations}
\alias{ghcn_read_stations}
\title{Read ghcnd stations file}
\usage{
ghcn_read_stations()
}
\value{
Data frame of stations data.
}
\description{
Read ghcnd stations file from the NOAA FTP site.
}
