% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{make_consistent}
\alias{make_consistent}
\title{Make unit of measurement consistent across variables}
\usage{
make_consistent(x, element)
}
\arguments{
\item{x}{Vector of values of type integer or double}

\item{element}{GHCN element type of size x. For example, "PRCP" is the
precipitation.}
}
\value{
Vector of type double.
}
\description{
Always try to use millimeters for variables that measure length, and degrees
Celsius for variables that measure temperature.
}
