% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dly_file}
\alias{read_dly_file}
\title{Read a ".dly" file}
\usage{
read_dly_file(file)
}
\arguments{
\item{file}{File path of ".dly" file.}
}
\value{
Data frame.
}
\description{
Read a ".dly" file, given the file path.
}
