% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-package.R
\docType{package}
\name{ghcn-package}
\alias{ghcn}
\alias{ghcn-package}
\title{ghcn: Tools for Working with GHCN Climate Data}
\description{
Tools for working with GHCN climate data. Read and
    clean daily climate data.
}
\author{
\strong{Maintainer}: Ray Wright \email{raywright88@gmail.com}

}
\keyword{internal}
