% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dly}
\alias{read_dly}
\title{Read a ".dly" file}
\usage{
read_dly(id)
}
\arguments{
\item{id}{Character vector of a station identification code.}
}
\value{
Data frame.
}
\description{
Read a ".dly" file from the NOAA FTP site, given the station identification
code.
}
